ai-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-genai==1.14.0
sentence-transformers==3.4.1
faiss-cpu==1.9.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Phase 2: Knowledge Base
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
httpx==0.28.0
beautifulsoup4==4.12.3
lxml==5.3.0
rank-bm25==0.2.2
apscheduler==3.10.4
tiktoken==0.8.0
bleach==6.2.0
tqdm==4.67.0
